+++ R/dataset.R
@@ -190,7 +190,7 @@ setMethod("refresh", "CrunchDataset", function (x) {
 #' for datasets and \code{\link{hide}} for variables.
 #'
 #' Deleting requires confirmation. In an interactive session, you will be asked
-#' to confirm. To avoid that prompt, or to delete objects from a 
+#' to confirm. To avoid that prompt, or to delete objects from a
 #' non-interactive session, wrap the call in \code{\link{with_consent}} to give
 #' your permission to delete.
 #'
@@ -414,9 +414,9 @@ publish <- function (x) {
 #' list of settings supported throughout the Crunch system. Clients may also
 #' provide and use custom settings if they choose.
 #' @param x CrunchDataset
-#' @param value A settings object (\code{ShojiEntity}), for the setter
-#' @return The getter returns a settings object (\code{ShojiEntity}). The setter
-#' returns the dataset (\code{x}).
+#' @param value A settings object (`ShojiEntity`), for the setter
+#' @return The getter returns a settings object (`ShojiEntity`). The setter
+#' returns the dataset (`x`).
 #' @examples
 #' \dontrun{
 #' settings(ds)
@@ -435,3 +435,36 @@ settings <- function (x) {
     updateEntity(settings(x), value)
     return(x)
 }
+
+#' View or set a dashboard URL
+#'
+#' You can designate a dashboard that will show when the dataset is loaded in
+#' the Crunch web app. This dashboard could be a Crunch Shiny ("Crunchy") app,
+#' a CrunchBox, or something else.
+#'
+#' @param x CrunchDataset
+#' @param value For the setter, a URL (character) or `NULL` to unset the
+#' dashboard.
+#' @return The getter returns a URL (character) or `NULL`. The setter
+#' returns the dataset (`x`).
+#' @examples
+#' \dontrun{
+#' dashboard(ds) <- "https://shiny.crunch.io/example/"
+#' }
+#' @export
+dashboard <- function (x) {
+    stopifnot(is.dataset(x))
+    app_settings <- x@body[["app_settings"]] %||% list()
+    whaam <- app_settings[["whaam"]] %||% list()
+    return(whaam[["dashboardUrl"]])
+}
+
+#' @rdname dashboard
+#' @export
+setDashboardURL <- function (x, value) {
+    setEntitySlot(x, "app_settings", list(whaam=list(dashboardUrl=value)))
+}
+
+#' @rdname dashboard
+#' @export
+"dashboard<-" <- setDashboardURL