+++ NEWS.md
@@ -1,4 +1,6 @@
 ### crunch 1.16.1 (under development)
+
+* `dashboard` and `dashboard<-` methods to view and set a dashboard URL on a dataset
 * Fix issue in printing filter expressions with long value columns
 
 ## crunch 1.16.0